Fix IDv1SN bug + renaming Indonesian layout sort problem (#438)
* Add Indonesian Keyboard v1 with Symbols and Number

The reason behind making these change is that i want to just insert the character without needing to click on "Pound Button" which is time consuming (for me). i added numeric number to 10 of the empty slot just like symbol and alphabet. Is this change possible?

Layout (i think) should be around like this
   |`^ |+++
 T2|+d!|?i3
$1h|/g\|m=+
-----------
+4+|xfc|+++
+r+|zsj|ke+
+5+|vwy|+++
-----------
+6p|+l+|o++
+n7|+uq|+a9
+8+|+++|+0+

I'm too lazy to write all the layout (the plus '+' are just redundant), but it should be around like this, i'm only add the numbers, the remains are untouched.

* repair from 'symbols to sn'

* Fix IDv1SN

There's a problem with IDv1SN regarding SwipeNWay usage. I removed it so it will behave correctly.

* Fix formatting

* Fix Capitalizing Indonesian Layout

Fixing by Capitalizing the name so it will be sorted correctly in alphabetical order.