Created layout for Canadian Aboriginal Syllabics (#472)
* Created layout for Canadian Aboriginal Syllabics

Explanation:
https://en.wikipedia.org/wiki/Canadian_Aboriginal_syllabics

/!\ Disclaimer: I don't speak any languages that use it, in fact I'm from the other side of the ocean. But I just think it's neat.

This actually relies on the vowel geometrics of this beautiful script and the swipe functions of this clever keyboard, to smartly encode a syllabary. Swiping the same direction always produces the same vowel!

Shift is mapped to long vocals, the palatalized forms and some of the Inuit forms are yet missing but I'm running out of keys, and don't know how to add a third layer without sacrificing the numerals.

Again this is untested, and just made by an amateur. I would love to try it out for real though!

* Removed some unnessesairy blank lines

* Update KeyboardLayout.kt with Canadian Aboriginal Syllabics

I hope I type this correctly, I'm doing this directly from my phone 😁

* Fixing lint.

---------

Co-authored-by: Dessalines <tyhou13@gmx.com>