PLTypeSplit*: Move a few letters to more sensible places
Letter positions in PLTypeSplit and PLTypeSplitSymbols layouts are based
on QWERTY, however a few letters reside in unexpected places, which
makes writing a bit unintuitive. The middle row is also unbalanced -
left half has 4 letters, while right side has 8 (granted, these are less
frequently used ones).

Move a few letters to more familiar positions:
* Move G to left half
* Move F to left half
* Move J more to the left
* Move ` and ~ elsewhere to make place for F and G (only
  PLTypeSplitSymbols)

Note that one letter is still in a "wrong" place: D, which should be in
the left half. However, it can't be moved as it's more popular than all
letters in the right side middle row (except K), and A and S are more
popular and they already occupy both available slots in the left half.

Signed-off-by: h67ma <szycikm@gmail.com>