NumericTypeSplit: reorganize interpunction and symbols
On *TypeSplit keyboards, the numeric keyboard is often the only way to
input most special symbols. However, those symbols are poorly organized.
The ones commonly found on the numeric row ('!', '@', '#', ...) are
found on seemingly random digit keys, making it hard to localize them.
Furthermore, the layout is cluttered, with many symbols packed on a
single key (in some cases up to 7 symbols on one key).

Move numeric row symbols to keys where the user would be expecting them:
'!' to 1, '@' to 2, '#' to 3, etc., with the exception of '(', which is
left on the side of the keyboard, so that the brackets lie in
symmetrical places. Each of the 8 symbols is activated by swiping
towards 5 (and '%' is swipe down).

Move a few symbols to one of the currently empty spaces on the left side
to declutter the numpad area.

Remove ',' and '.' from numpad, as they are already available on the
right side of the layout.

Move a few other other symbols to be more evenly spaced out and easier
to access.

Signed-off-by: h67ma <szycikm@gmail.com>