Add some swipeReturn actions to the french ThumbKey v2 layout
In addition, I moved the æ to the top left to move @ at the same place
as in the numeric layout.
This is not in the relative same place than œ, but I think it’s ok.
This gave me a free space to put ) on top-right of A, and a ( on
top-left of R, which is usefull when typing.

For the swipeReturn, I added the following :

* [ ] on the added ( ) keys ;
* { } on the already existing « and » ;
* ; on the , and : on the . ;
* \ on the numeric layer /, allowing it to be used on the main layer
  with the recent patch #1158 of @gitterrost4 ;
* " on the numeric layer ', for the same reason as for \.

This is working now, and will work better with the recents patch #1156
from @gitterrost4.