fix numeric wide underscore
The cause of the bug was that `swipeType` was set
to `FOUR_WAY_DIAGONAL` instead of `EIGHT_WAY`.
However, instead of changing the `swipeType`,
I decided to adopt @a6-webm 's solution and shift
the underscore to `bottomLeft` so that it the key
positions matches the rest of the this wide
numeric layout.

fixes #1597