Restore MessagEase compatibility for deutsch + æøå
In pull request #832, merged as commit 77e058f, the layout
"deutsch+æøå" was introduced for compatibility to the similarily named
layout "German+ÆØÅ" in MessagEase.

In the commit message of that change, it was explicitly noted the
differences between this and the "deutsch" layout, notably:

(2) Top swipe on center left key H is Æ instead of Ü
(4) Top swipe on lower left key in numeric keyboard is umlaut/diaeresis
(5) Bottom swipe on lower left key T does not show eszett

In issue #1027/pull request #1028, merged as commit 5458d36, user
@MarvinBaral undid these points. No comment was solicited from the
author of the original, nor was there any discussion of these changes.

In the issue it was simply stated that the characters ü and ß was
"missing" and that their omission is a "mistake".

However, these differences were quite intentional, as the variant is
not intended to be German with some funny characters added. Rather,
its entire rationale is that the Scandinavian branch of North-Germanic
languages (i.e. Norwegian, Swedish, Danish) have a lot of phonology in
common with German, but with some distinct differences. In particular
ü is used only in foreign proper names, and ß is never
used. Conversely, æ, ø and å is heavily used in Scandinavian but not
present as characters in regular German at all. If there is any error,
it is in not naming the variant "Scandinavian" or "Nordic" instead.

In this light, the changes in #1028 seems scantly motivated, as they
changed the "deutsch + æøå" variant into something for which there is
little demand other than perhaps allowing the author to write both
Norwegian and German with the same layout.

A lot of research went into the MessagEase layouts to make sure that
phonemes should be easily tapped in the target languages. I think that
the bar for breaking this should be higher than a single user's
personal preferences, in particular one that changes the layout for
everyone else.

This patch changes the layout back to as it was originally submitted.