Add MessagEase "German + ÅØÆ" layout as "scandinavian"
This patch adds the layout that was known as "German + ÅØÆ" in
MessagEase into Thumb-Key under the name "scandinavian", as it is
intended to serve the modern standard languages Norwegian, Swedish and
Danish of the North-Germanic branch.

See #832, merged in commit 77e058f, for a precursor of this layout,
and an explanation of the differences between it and the "deutsch"
layout.