Updates to adhere to naming convention
The first version did not have a base portion "messagease" because
although the layout has its origins in a layout that was bundled with
MessagEase, there was never anyone called "Scandinavian" in that
product.

It is named Scandinavian instead of for instance Norwegian, Swedish or
Danish because, as all these languages are mutually intelligible to
the degree that they are more dialects with each their own navy, the
original MessagEase layout was not targeting any particular one of
them but rather all of them as a common group.